moved caveminer files
