+++ map_layout.lua
@@ -19,7 +19,7 @@ local tiles_per_tick = 32
 --require "map_gen.combined.dimensions"
 --require "map_gen.combined.dagobah_swamp"
 --require "map_gen.combined.meteor_strike" --unfinished
---require 'cave_miner'
+--require 'map_gen/combined/cave_miner/cave_miner'
 
 
 --presets--
